refactor: Move http utils functions to httputils.go file
