+++ datacall.go
@@ -2,16 +2,14 @@ package main
 
 import (
 	"encoding/json"
-	"io/ioutil"
-	"net/http"
 	"time"
 )
 
 // KotoData is DTO of thing to do in Yarana-Bot
 type KotoData struct {
-	ID     string
-	UserID string
-	Title  string
+	ID     string `json:"id"`
+	UserID string `json:"userId"`
+	Title  string `json:"title"`
 }
 
 // NewKotoData is constructor of KotoData
@@ -65,8 +63,6 @@ type DataCall interface {
 type YaranaDataCall struct {
 }
 
-// TODO: Implement methods of YaranaDataCall
-
 // NewYaranaDataCall is a constructor of YaranaDataCall
 func NewYaranaDataCall() (*YaranaDataCall, error) {
 	return &YaranaDataCall{}, nil
@@ -84,41 +80,6 @@ func (c *YaranaDataCall) GetKotosByUserID(userID string) (kotos []*KotoData, err
 		return nil, err
 	}
 	return kotos, nil
-
-	/* Mock implementation
-	// Get Koto by userID from something
-	id := "0123456789a"
-	title := "Test Title"
-	koto, err := NewKotoData(id, userID, title)
-	if err != nil {
-		return nil, err
-	}
-	title2 := "Test Title 2"
-	koto2, err := NewKotoData(id, userID, title2)
-	if err != nil {
-		return nil, err
-	}
-	return []*KotoData{koto, koto2}, nil
-	*/
-}
-
-// AssembleURLWithParam makes url with parameter
-func AssembleURLWithParam(baseURL string, key string, val string) string {
-	return baseURL + "?" + key + "=" + val
-}
-
-// HTTPGet curls with arg url
-func HTTPGet(url string) (buf []byte, err error) {
-	res, err := http.Get(url)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-	buf, err = ioutil.ReadAll(res.Body)
-	if err != nil {
-		return nil, err
-	}
-	return buf, err
 }
 
 // AddKoto is a method of DataCall interface
